Use mean for currently open tabs

+++ src/main/scala/statistics/StatisticsOutput.scala
@@ -1,14 +1,14 @@
 package statistics
 
 case class StatisticsOutput(
-    numCurrentTabs: Float,
-    openTabsGrouped: Float,
-    openTabsUngrouped: Float,
-    tabSwitchWithinGroups: Float,
-    tabSwitchBetweenGroups: Float,
-    tabSwitchFromGroup: Float,
-    tabSwitchToGroup: Float,
-    tabSwitchUngrouped: Float
+    numCurrentTabs: Double,
+    openTabsGrouped: Double,
+    openTabsUngrouped: Double,
+    tabSwitchWithinGroups: Int,
+    tabSwitchBetweenGroups: Int,
+    tabSwitchFromGroup: Int,
+    tabSwitchToGroup: Int,
+    tabSwitchUngrouped: Int
 ) {
   def asCsv: String = {
     s"$numCurrentTabs;$openTabsGrouped;$openTabsUngrouped;$tabSwitchWithinGroups;$tabSwitchBetweenGroups;$tabSwitchFromGroup;$tabSwitchToGroup;$tabSwitchUngrouped"
